Add absorber factory and move implementations to internal directories
